fix: fix naming

+++ src/Lab1/Environment/Models/HighDensityFog.cs
@@ -2,7 +2,7 @@ namespace Itmo.ObjectOrientedProgramming.Lab1.Environment.Models;
 
 public class HighDensityFog : Entities.Environment
 {
-    public HighDensityFog(int antimatterFlaresCount, int lengthOfEnvironment)
+    public HighDensityFog(int antimatterFlaresCount, double lengthOfEnvironment)
     {
         AddRequiredEngineName("Impulse Engine Class C");
         AddRequiredEngineName("Impulse Engine Class E");
