+++ src/Lab1/Environment/Models/NitrineParticleFog.cs
@@ -2,7 +2,7 @@ namespace Itmo.ObjectOrientedProgramming.Lab1.Environment.Models;
 
 public class NitrineParticleFog : Entities.Environment
 {
-    public NitrineParticleFog(int spaceWhalesCount, int lengthOfEnvironment)
+    public NitrineParticleFog(int spaceWhalesCount, double lengthOfEnvironment)
     {
         AddRequiredEngineName("Impulse Engine Class E");
         AddObstacle("SpaceWhale", spaceWhalesCount);
