+++ tests/Lab1.Tests/FourthTest.cs
@@ -9,8 +9,8 @@ namespace Itmo.ObjectOrientedProgramming.Lab1.Tests;
 public class FourthTest
 {
     [Theory]
-    [InlineData(10000, 10000, 12, 14, (int)DistanceType.Small)]
-    public void TestRouteForSpaceships(double initialFuelActivePlasma, double initialFuelGravitonMatter, double fuelActivePlasmaPrice, double fuelGravitonMatterPrice, int normalSpaceLength)
+    [InlineData(10000, 10000, 12, 14, (double)DistanceType.Small)]
+    public void TestRouteForSpaceships(double initialFuelActivePlasma, double initialFuelGravitonMatter, double fuelActivePlasmaPrice, double fuelGravitonMatterPrice, double normalSpaceLength)
     {
         // Arrange
         var spaceship1 = new PleasureShuttleShip();
