+++ tests/Lab1.Tests/SeventhTest.cs
@@ -9,8 +9,8 @@ namespace Itmo.ObjectOrientedProgramming.Lab1.Tests;
 public class SeventhTest
 {
     [Theory]
-    [InlineData(10000, 10000, 1, 1, (int)DistanceType.Big, (int)DistanceType.Small, (int)DistanceType.Medium)]
-    public void TestRouteForSpaceships(double initialFuelActivePlasma, double initialFuelGravitonMatter, double fuelActivePlasmaPrice, double fuelGravitonMatterPrice, int normalSpaceLength, int highDensityFogLength, int nitrineParticleFogLength)
+    [InlineData(10000, 10000, 1, 1, (double)DistanceType.Big, (double)DistanceType.Small, (double)DistanceType.Medium)]
+    public void TestRouteForSpaceships(double initialFuelActivePlasma, double initialFuelGravitonMatter, double fuelActivePlasmaPrice, double fuelGravitonMatterPrice, double normalSpaceLength, double highDensityFogLength, double nitrineParticleFogLength)
     {
         // Arrange
         var spaceship1 = new PleasureShuttleShip();
